Landing page separation of concerns
